requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
boto3
langchain-community
langchain_chroma
bs4
unstructured
markdown
python-dotenv
openai
langgraph
ollama
langchain_openai
